Added Working Pause Menu
There is now a pause menu accessible by pressing escape. The main menu
at the moment is a temporary way to get back to main menu. Plan to
change that as more of end game call and send to end game screen. Resume
button is there if player doesn't want to hit escape again to unpause.